+++ cmd/geth/usage.go
@@ -144,9 +144,7 @@ var AppHelpFlagGroups = []flagGroup{
 	{
 		Name: "VIRTUAL MACHINE",
 		Flags: []cli.Flag{
-			utils.VMEnableJitFlag,
-			utils.VMForceJitFlag,
-			utils.VMJitCacheFlag,
+			utils.VMCacheFlag,
 		},
 	},
 	{
