+++ core/vm/program_test.go
