+++ core/vm/program_util.go
