+++ core/vm/program_util_test.go
