+++ core/vm_env.go
@@ -41,11 +41,11 @@ func GetHashFn(ref common.Hash, chain *BlockChain) func(n uint64) common.Hash {
 }
 
 type VMEnv struct {
-	chainConfig *ChainConfig   // Chain configuration
-	state       *state.StateDB // State to use for executing
-	evm         *vm.EVM        // The Ethereum Virtual Machine
-	depth       int            // Current execution depth
-	msg         Message        // Message appliod
+	chainConfig *ChainConfig      // Chain configuration
+	state       *state.StateDB    // State to use for executing
+	evm         vm.VirtualMachine // The Ethereum Virtual Machine
+	depth       int               // Current execution depth
+	msg         Message           // Message appliod
 
 	header    *types.Header            // Header information
 	chain     *BlockChain              // Blockchain handle
